+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.1
+- Added a button to open the BepInEx folder
+- Change Centered the buttons in the UI
+- Renamed the "Install Modloader" button to "Install BepInEx"
+- Renamed the "Uninstall Modloader" button to "Uninstall BepInEx"
+- Removed Melon Loader
+
 # 1.0.0
 - Added a button to install BepInEx and MelonLoader
 - Added a button to uninstall BepInEx and MelonLoader